+++ CHANGELOG.md
@@ -238,4 +238,5 @@
 * Made several methods of PlayerShieldComponent public
 * Fixed IGUI tabs added after the fact not receiving themeing
 * Fixed Button text going invisible when element color class is overridden to None
+* Exposed a few state fields in PlayerWeaponComponent as read only properties
 * Fixed ThemeableElement and NonThemeableElement being ignored on child components of complex elements [WIP]